move switch layout code into FieldTemplate and convert to SwitchButtonWidget
