+++ src/components/form/widgets/switchButton/SwitchButtonWidget.tsx
@@ -15,26 +15,22 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import React, { ChangeEvent, ReactNode } from "react";
-import { Col, Form as BootstrapForm, Row } from "react-bootstrap";
+import React, { ChangeEvent } from "react";
 import BootstrapSwitchButton from "bootstrap-switch-button-react";
-import styles from "./SwitchButton.module.scss";
+import { Except } from "type-fest";
+import { FieldProps } from "@/components/form/FieldTemplate";
 
 export type CheckBoxLike = {
   name: string;
   value: boolean;
 };
 
-export type SwitchButtonProps = {
-  name: string;
-  label: ReactNode;
+export type SwitchButtonProps = Except<FieldProps, "onChange"> & {
   onChange: React.ChangeEventHandler<CheckBoxLike>;
-  value?: boolean;
 };
 
-const SwitchButton: React.FC<SwitchButtonProps> = ({
+const SwitchButtonWidget: React.FC<SwitchButtonProps> = ({
   name,
-  label,
   onChange,
   value,
 }) => {
@@ -45,20 +41,13 @@ const SwitchButton: React.FC<SwitchButtonProps> = ({
   };
 
   return (
-    <BootstrapForm.Group as={Row} controlId={name}>
-      <Col sm="3">
-        <BootstrapSwitchButton
-          onlabel=" "
-          offlabel=" "
-          checked={value}
-          onChange={patchedOnChange}
-        />
-      </Col>
-      <Col sm="9" className={styles.label}>
-        {label}
-      </Col>
-    </BootstrapForm.Group>
+    <BootstrapSwitchButton
+      onlabel=" "
+      offlabel=" "
+      checked={value}
+      onChange={patchedOnChange}
+    />
   );
 };
 
-export default SwitchButton;
+export default SwitchButtonWidget;
