+++ src/components/formBuilder/FieldEditor.tsx
@@ -33,7 +33,8 @@ import FieldTemplate from "@/components/form/FieldTemplate";
 import { produce } from "immer";
 import SelectWidget from "@/components/form/widgets/SelectWidget";
 import OptionsWidget from "@/components/form/widgets/OptionsWidget";
-import { CheckBoxLike } from "@/components/form/switchButton/SwitchButton";
+import SwitchButtonWidget from "@/components/form/widgets/switchButton/SwitchButtonWidget";
+import { CheckBoxLike } from "@/components/form/widgets/switchButton/SwitchButtonWidget";
 import { uniq } from "lodash";
 
 const FieldEditor: React.FC<{
@@ -267,7 +268,7 @@ const FieldEditor: React.FC<{
       <FieldTemplate
         name={`${name}.schema.required`}
         label="Required Field?"
-        layout="switch"
+        as={SwitchButtonWidget}
         value={isRequired}
         onChange={onRequiredChange}
       />
